Add span-based SQL templates
